+++ pwnlib/adb/__init__.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 from pwnlib.adb.adb import *
 from pwnlib.adb.protocol import Client
 from pwnlib.adb.bootimg import BootImage