+++ pwnlib/encoders/arm/xor.py
@@ -1,6 +1,3 @@
-from __future__ import absolute_import
-from __future__ import division
-
 from pwnlib import shellcraft
 from pwnlib.asm import asm
 from pwnlib.context import context